Make ArchiveTrieNodeStrategy a InSyncSubscriber
